pyproject.toml
[project]
name = "kycflow"
version = "0.1.0"
description = "Project made for kyc verification using Yunet + InsightFace"
authors = [
    { name = "AdityaSinghDevs", email = "adityasingh.devs@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10,<3.14"

package-mode = false

dependencies = [
    "pyyaml>=6.0.3,<7.0.0",
    "pydantic>=2.11.9,<3.0.0",
    "fastapi>=0.118.0,<0.119.0",
    "uvicorn>=0.37.0,<0.38.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "torch>=2.8.0,<3.0.0",
    "torchvision>=0.23.0,<0.24.0",
    "insightface>=0.7.3,<0.8.0",
    "easyocr>=1.7.2,<2.0.0",
    "opencv-python-headless>=4.12.0.88,<5.0.0.0",
    "pillow>=11.3.0,<12.0.0",
    "requests>=2.32.5,<3.0.0",
    "tqdm>=4.67.1,<5.0.0.0",
    "pydantic-settings (>=2.11.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.4.2,<9.0.0"
pytest-asyncio = ">=1.2.0,<2.0.0"
black = ">=25.9.0,<26.0.0"
isort = ">=6.1.0,<7.0.0"
python-multipart = "^0.0.20"
